requirements.txt
streamlit>=1.44,<2
pandas>=2.2,<3
python-dotenv>=1.0,<2
pytest>=8.0,<9
ruff>=0.11,<1
